vanish check immediately following online check
move the check to see if a player is trying to trade with a player that is vanished to be right after the check if the player is offline. this resolves the issue caused where a player could use /trade <vanished-player> and if the vanished player was in a non-supported gamemode it would tell them that instead of saying they are offline. (also other checks in between but this is the one that made me realize this was an issue)